+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-06T18:15:00.546Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-07T11:47:14.495Z")
     r"""The date and time the resource was last updated"""