+++ packages/orq-rc/RELEASES.md
@@ -679,3 +679,13 @@ Based on:
 - [python v3.2.0-rc.30] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.30] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.30 - packages/orq-rc
+
+## 2025-03-07 11:48:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.513.2 (2.545.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.31] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.31] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.31 - packages/orq-rc